refactor(tests/commands): separate command unit tests into modules
