+++ tests/commands/conftest.py
@@ -1,0 +1,11 @@
+import pytest
+
+from commitizen import defaults
+from commitizen.config import BaseConfig
+
+
+@pytest.fixture()
+def config():
+    _config = BaseConfig()
+    _config.settings.update({"name": defaults.name})
+    return _config
