+++ tests/commands/test_check_command.py
@@ -1,8 +1,22 @@
 import sys
+import contextlib
+import os
+import shutil
+import tempfile
 
 import pytest
 
 from commitizen import cli
+from commitizen import commands
+
+
+@contextlib.contextmanager
+def get_temp_dir():
+    temp_dir = tempfile.mkdtemp()
+    try:
+        yield temp_dir
+    finally:
+        shutil.rmtree(temp_dir)
 
 
 def test_check_jira_fails(mocker, capsys):
@@ -76,3 +90,41 @@ def test_check_conventional_commit_succeeds(mocker, capsys):
     cli.main()
     out, _ = capsys.readouterr()
     assert "Commit validation: successful!" in out
+
+
+def test_check_no_conventional_commit(config, mocker, tmpdir):
+    with pytest.raises(SystemExit):
+        error_mock = mocker.patch("commitizen.out.error")
+
+        with get_temp_dir() as dir:
+
+            tempfile = os.path.join(dir, "temp_commit_file")
+            with open(tempfile, "w") as f:
+                f.write("no conventional commit")
+
+            check_cmd = commands.Check(
+                config=config, arguments={"commit_msg_file": tempfile}
+            )
+            check_cmd()
+            error_mock.assert_called_once()
+
+
+def test_check_conventional_commit(config, mocker):
+    success_mock = mocker.patch("commitizen.out.success")
+    with get_temp_dir() as dir:
+
+        tempfile = os.path.join(dir, "temp_commit_file")
+        with open(tempfile, "w") as f:
+            f.write("feat(lang): added polish language")
+
+        check_cmd = commands.Check(
+            config=config, arguments={"commit_msg_file": tempfile}
+        )
+
+        check_cmd()
+        success_mock.assert_called_once()
+
+
+def test_check_command_when_commit_file_not_found(config):
+    with pytest.raises(FileNotFoundError):
+        commands.Check(config=config, arguments={"commit_msg_file": ""})()
